Extract ScrobblerProxy to own file
